Trying to fix Serialization issues - ongoing
