+++ Assets/Scripts/FileLeaderboard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab51bb3a821ec9c2f81256ee99f3cdac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
